Refactor IM commands
Move general commands from IM commands to main commands.
