+++ cypress/utils/plugins/index-management-dashboards-plugin/commands.js
@@ -3,40 +3,41 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { BACKEND_BASE, IM_API, IM_CONFIG_INDEX } from '../../constants';
+import { BACKEND_BASE_PATH, IM_API, IM_CONFIG_INDEX } from '../../constants';
 
-Cypress.Commands.add('deleteAllIndices', () => {
-  cy.request(
-    'DELETE',
-    `${Cypress.env('openSearchUrl')}/index*,sample*,opensearch_dashboards*`
-  );
+Cypress.Commands.add('deleteIMJobs', () => {
   // TODO don't directly delete system index, use other way
   cy.request(
     'DELETE',
     `${Cypress.env('openSearchUrl')}/.opendistro-ism*?expand_wildcards=all`
   );
   // Clean all ISM policies
-  cy.request('GET', `${BACKEND_BASE}${IM_API.POLICY_BASE}`).then((resp) => {
-    const policyIds = resp.body.policies.map((item) => item._id).join(', ');
-    if (policyIds.length) {
-      cy.request('DELETE', `${BACKEND_BASE}${IM_API.POLICY_BASE}/${policyIds}`);
+  cy.request('GET', `${BACKEND_BASE_PATH}${IM_API.POLICY_BASE}`).then(
+    (resp) => {
+      const policyIds = resp.body.policies.map((item) => item._id).join(', ');
+      if (policyIds.length) {
+        cy.request(
+          'DELETE',
+          `${BACKEND_BASE_PATH}${IM_API.POLICY_BASE}/${policyIds}`
+        );
+      }
     }
-  });
+  );
   // Clean all ISM jobs
-  cy.request('GET', `${BACKEND_BASE}${IM_API.EXPLAIN_BASE}`).then((resp) => {
-    console.log('Explain all');
-    console.log(`${JSON.stringify(resp.body)}`);
-    let ismJobIds = Object.keys(resp.body);
-    ismJobIds = ismJobIds
-      .filter((e) => e !== 'total_managed_indices')
-      .join(', ');
-    if (ismJobIds.length) {
-      cy.request(
-        'POST',
-        `${BACKEND_BASE}${IM_API.REMOVE_POLICY_BASE}/${ismJobIds}`
-      );
+  cy.request('GET', `${BACKEND_BASE_PATH}${IM_API.EXPLAIN_BASE}`).then(
+    (resp) => {
+      let ismJobIds = Object.keys(resp.body);
+      ismJobIds = ismJobIds
+        .filter((e) => e !== 'total_managed_indices')
+        .join(', ');
+      if (ismJobIds.length) {
+        cy.request(
+          'POST',
+          `${BACKEND_BASE_PATH}${IM_API.REMOVE_POLICY_BASE}/${ismJobIds}`
+        );
+      }
     }
-  });
+  );
 });
 
 Cypress.Commands.add('createPolicy', (policyId, policyJSON) => {
@@ -44,18 +45,6 @@ Cypress.Commands.add('createPolicy', (policyId, policyJSON) => {
     'PUT',
     `${Cypress.env('openSearchUrl')}${IM_API.POLICY_BASE}/${policyId}`,
     policyJSON
-  );
-});
-
-Cypress.Commands.add('getIndexSettings', (index) => {
-  cy.request('GET', `${Cypress.env('openSearchUrl')}/${index}/_settings`);
-});
-
-Cypress.Commands.add('updateIndexSettings', (index, settings) => {
-  cy.request(
-    'PUT',
-    `${Cypress.env('openSearchUrl')}/${index}/_settings`,
-    settings
   );
 });
 
@@ -85,50 +74,12 @@ Cypress.Commands.add('updateManagedIndexConfigStartTime', (index) => {
   });
 });
 
-Cypress.Commands.add('createIndex', (index, policyID = null, settings = {}) => {
-  cy.request('PUT', `${Cypress.env('openSearchUrl')}/${index}`, settings);
-  if (policyID != null) {
-    const body = { policy_id: policyID };
-    cy.request(
-      'POST',
-      `${Cypress.env('openSearchUrl')}${IM_API.ADD_POLICY_BASE}/${index}`,
-      body
-    );
-  }
-});
-
 Cypress.Commands.add('createRollup', (rollupId, rollupJSON) => {
   cy.request(
     'PUT',
     `${Cypress.env('openSearchUrl')}${IM_API.ROLLUP_JOBS_BASE}/${rollupId}`,
     rollupJSON
   );
-});
-
-Cypress.Commands.add('createIndexTemplate', (name, template) => {
-  cy.request(
-    'PUT',
-    `${Cypress.env('openSearchUrl')}${IM_API.INDEX_TEMPLATE_BASE}/${name}`,
-    template
-  );
-});
-
-Cypress.Commands.add('createDataStream', (name) => {
-  cy.request(
-    'PUT',
-    `${Cypress.env('openSearchUrl')}${IM_API.DATA_STREAM_BASE}/${name}`
-  );
-});
-
-Cypress.Commands.add('deleteDataStreams', (names) => {
-  cy.request(
-    'DELETE',
-    `${Cypress.env('openSearchUrl')}${IM_API.DATA_STREAM_BASE}/${names}`
-  );
-});
-
-Cypress.Commands.add('rollover', (target) => {
-  cy.request('POST', `${Cypress.env('openSearchUrl')}/${target}/_rollover`);
 });
 
 Cypress.Commands.add('createTransform', (transformId, transformJSON) => {
